requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
pytest==8.4.1
python-dotenv==1.1.1
selene==2.0.0rc9
selenium==4.35.0
pytest-xdist==3.8.0
